THIRDEYE-308 High-loss of dimensional data (most records go to other) from Kafka consumer
There was an issue in the DimensionStoreImmutableImpl code in which we were
picking the first record to match, which is not always the record with least
others. This fixes that and adds a test case for it.

However, the issue seen with ads is a discrepancy between the star tree used
and the data in Kafka. For example, a leaf node has the following combinations:

[2, 11, ?, *,  ]
[2, 11, ela4, *,  ]
[2, 11, ltx1, *,  ]
[2, 11, lva1, *,  ]
[2, 14, ?, *,  ]
[2, 14, ?, *, ?]
[2, 14, ela4, *,  ]
[2, 14, ltx1, *,  ]
[2, 14, lva1, *,  ]
[2, 15, ?, *,  ]
[2, 15, ela4, *,  ]
[2, 15, ltx1, *,  ]
[2, 15, lva1, *,  ]
[2, 5, ?, *,  ]
[2, 5, ela4, *,  ]
[2, 5, ltx1, *,  ]
[2, 5, lva1, *,  ]
[2, ?, ?, *, ?]

And we are trying to insert [2, 14, lva1, *, unknown].

This maps to [2, 14, ?, *, ?], so we lose the colo name (which is one of the
weird issues on the dashboard), but this is because we did not see the
"unknown" os name in the training set (it was " ") instead.

So with this fix, we can probably proceed (working with the engagement) to
generate a new tree that corresponds to the data in Kafka.

RB=455065
R=kgopalak,npawar
A=npawar